+++ app/api/auth/[...nextauth]/route.ts
@@ -1,15 +1,5 @@
 import NextAuth from "next-auth";
-import GithubProvider from "next-auth/providers/github";
-import type { NextAuthOptions } from "next-auth";
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID!,
-      clientSecret: process.env.GITHUB_SECRET!,
-    }),
-  ],
-};
+import { authOptions } from "@/lib/auth";
 
 const handler = NextAuth(authOptions);
