+++ lib/auth.ts
@@ -1,0 +1,12 @@
+import GithubProvider from "next-auth/providers/github";
+import type { NextAuthOptions } from "next-auth";
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    GithubProvider({
+      clientId: process.env.GITHUB_ID!,
+      clientSecret: process.env.GITHUB_SECRET!,
+      // Add redirectUri so can use diff one locally to in prod
+    }),
+  ],
+};
